add fields verification

+++ app/src/main/java/com/zerox/ticketmanager/data/model/Repository.kt
@@ -21,7 +21,7 @@ class Repository @Inject constructor(private val userDao: UserDao) {
         return user
     }
 
-    suspend fun addUser(user:UserEntitiy):Int{
+    suspend fun addUser(user:UserEntitiy):Long{
         return userDao.insertUser(user)
     }
 }
