+++ app/src/main/java/com/zerox/ticketmanager/data/model/database/dao/UserDao.kt
@@ -16,5 +16,5 @@ interface UserDao {
     suspend fun getUserByUsername(username:String):UserEntitiy
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
-    suspend fun insertUser(user : UserEntitiy):Int
+    suspend fun insertUser(user : UserEntitiy):Long
 }
